redirect

+++ project/apps/about/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 
 class CallBack(models.Model):
-    number = models.IntegerField(verbose_name='Телефон')
+    number = models.BigIntegerField(verbose_name='Телефон')
     name = models.CharField(max_length=40, verbose_name='Имя')
     created = models.DateTimeField(auto_now_add=True, verbose_name='Создано')
